Move search to separate file
