+++ src/mds_array_manipulation/__init__.py
@@ -1,4 +1,2 @@
-# read version from installed package
-from importlib.metadata import version
-# __version__ = version("mds_array_manipulation")
+# Read version from installed package
 __version__ = "0.0.1"
