+++ src/mds_array_manipulation/search_array.py
@@ -1,0 +1,33 @@
+def search_array(arr, elem):
+    """
+    Search for elem in a numpy array.
+    
+    Parameters
+    ----------
+    arr : numpy.array
+        The numpy array to search
+    elem : object to search for in the array
+    
+    Returns
+    -------
+    int
+        Index of the first occurence of the element in the array, -1 if not found
+        
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from mds_array_manipulation.mds_array_manipulation import search_array
+    >>> arr = np.array([20, 10, 20, 30, 50, 90, 60])
+    >>> search_array(arr, 50)
+    4
+    >>> search_array(arr, 100)
+    -1
+    >>> search_array(arr, 20)
+    0
+    """
+    i = 0 
+    while i < len(arr):
+        if arr[i] == elem:
+            return i
+        i += 1
+    return -1
