--- tests/test_mds_array_manipulation.py
@@ -1,1 +1,0 @@
-from src.mds_array_manipulation import mds_array_manipulation
